getting data from device and showing it

+++ lib/models/Temperature.dart
@@ -1,6 +1,6 @@
 class Temperature{
   final DateTime time;
-  final int temperature;
+  final double temperature;
 
   Temperature({required this.temperature, required this.time});
